% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthRaw.R
\name{monthRaw}
\alias{monthRaw}
\title{Subset large dataset to user-specified year and month}
\usage{
monthRaw(year, plants = FALSE, month)
}
\arguments{
\item{year}{Single year,from 2021 onwards}

\item{plants}{If not animals, select plants=TRUE}

\item{month}{Select by number 1 - 12}
}
\value{
a dataframe
}
\description{
Subset large dataset to user-specified year and month
}
\examples{
#monthRaw(2024, plants=FALSE, month=3)
}
