% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetMonths_Cust.R
\name{SubsetMonths_Cust}
\alias{SubsetMonths_Cust}
\title{Subset large dataset to user-specified year and month}
\usage{
SubsetMonths_Cust(year, plants = FALSE, month)
}
\arguments{
\item{year}{e.g. 2024}

\item{plants}{If not animals, select plants=TRUE}

\item{month}{Select by number 1 - 12}
}
\value{
a dataframe
}
\description{
Subset large dataset to user-specified year and month
}
\examples{
#SubsetMonths_Cust(2024, plants=FALSE, month=3)
}
