% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulateCommod_Cust.R
\name{TabulateCommod_Cust}
\alias{TabulateCommod_Cust}
\title{Generate a table of volume by type, month, location and origin for user-specifed year}
\usage{
TabulateCommod_Cust(year)
}
\arguments{
\item{year}{e.g. 2024}
}
\value{
a tibble
}
\description{
Generate a table of volume by type, month, location and origin for user-specifed year
}
\examples{
TabulateCommod_Cust(2024)
}
