% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearSummary.R
\name{yearSummary}
\alias{yearSummary}
\title{Generate a table of volume by type, month, location and origin for user-specifed year}
\usage{
yearSummary(year)
}
\arguments{
\item{year}{Single year,from 2021 onwards}
}
\value{
a tibble
}
\description{
Generate a table of volume by type, month, location and origin for user-specifed year
}
\examples{
#yearSummary(2024)
}
